Remove WordPoints components
WordPoints components are mainly internal to core, and this may be enforced in the future. As such, if they need to be dealt with here at all, they should probably be handled differently, perhaps under `wordpoints` info. At any rate, I don't think they are useful for the initial version of the spec.