Remove == version operator
Because version combinations are always assumed to be `AND` and not `OR`, the `==` operator is not very useful. It could not be used with any other operators, and using it alone would tie a dependent to a specific dependency version, which is not a good idea. As such, it seems best to remove it. In a pinch, the same effect can be achieved by combining the `>` and `<` operators.